requirements.txt
Flask
scikit-learn
python-docx
PyPDF2
